+++ server/app/model/resources/resources.js
@@ -229,7 +229,6 @@ ResourceSchema.statics.updateResourceById = function(resourceId, fields, callbac
 
 
 ResourceSchema.statics.getResourcesWithPagination=function(dataBaseQueryObj,callback){
-    dataBaseQueryObj.queryObj.isDeleted = false;
     Resources.paginate(dataBaseQueryObj.queryObj, dataBaseQueryObj.options, function(err, data) {
         if (err) {
             logger.error("Failed to getResourcesWithPagination", err);
